libpower: Replace private platform API with OSS HookMediator
First up, com.android.internal.telephony from telephony-common is, as
the name suggests, internal. It cannot be built for devices with an API
level higher than 28 where these private APIs are off-limits for vendor
applications.

While it would be possible to replace this with TelephonyManager which
is an allowed class, the codepaths for OemHooks are deprecated and
hidden. These rely on android.hardware.radio.deprecated::IOemHook which
is scheduled to disappear "soon".

All our devices will (be going to) run qcrild, that does not host this
deprecated interface anymore. Instead it has been replaced with
IQtiOemHook, a very similar interface that unfortunately does not
address the intial reasons for deprecating IOemHook.
While it is definitely possible (and easy) to get hold of an instance of
IQtiOemHook we need multiple applications doing so, which is impossible
and requires an intermediary service to relay commands and do the
accounting.

This change implements a connection to the HookMediatorService and calls
the transpower hook on it.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>